Fix output of sensor values
This is to fix an uninitialized value failure when looping through
oid_Sensors to print the temperature values. The error occur when the
device being polled does not have one or more sensors available in
oid_Sensors, thus the value will be uninitialized.
It makes more sense and is much safer to only print the output of
returned values instead.